Rename file to keep with convenention
