+++ tests/extensions/plugins/test_plugins.py
@@ -2,8 +2,8 @@ import copy
 
 from saleor.extensions import ConfigurationTypeField
 from saleor.extensions.models import PluginConfiguration
-from tests.extensions.helpers import get_config_value
 from tests.extensions.sample_plugins import PluginSample
+from tests.extensions.utils import get_config_value
 
 
 def test_update_config_items_keeps_bool_value(plugin_configuration):
